+++ Magic/CHANGELOG.md
@@ -69,6 +69,7 @@
  - Add `entity_parameters` section to SpawnEntity action, will be applied to any mobs spawned
  - Add `skills.quick_cast: auto` option to wands, to allow spellbooks with only some quick-cast skills
  - Add `rarify` spell as an optional Apprentice spell, improves ore drops but requires teamwork
+ - Magic attribute modifiers now support specific slots and operations like vanilla modifiers
 
 ## Changes